added security check

+++ src/main/java/ru/khanin/dmitrii/schedule/controller/FlowController.java
@@ -4,12 +4,14 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -18,30 +20,40 @@ import ru.khanin.dmitrii.schedule.dto.flow.FlowRequest;
 import ru.khanin.dmitrii.schedule.dto.flow.FlowResponse;
 import ru.khanin.dmitrii.schedule.entity.Flow;
 import ru.khanin.dmitrii.schedule.service.FlowService;
+import ru.khanin.dmitrii.schedule.service.UserService;
 
 @RestController()
 @RequestMapping("flow")
 @RequiredArgsConstructor
 public class FlowController {
 	private final FlowService flowService;
+	private final UserService userService;
 
 	@GetMapping("/all")
 	public ResponseEntity<List<FlowResponse>> getAllFlows() {
 		Collection<Flow> found = flowService.findAll();
 		List<FlowResponse> result = new ArrayList<>();
-		found.forEach((e) -> result.add(new FlowResponse(e.getFlowLvl(), e.getCourse(), e.getFlow(), e.getSubgroup())));
+		found.forEach((e) -> {
+			result.add(new FlowResponse(e.getFlowLvl(), e.getCourse(), e.getFlow(), e.getSubgroup()));
+		});
 		return ResponseEntity.ok(result);
 	}
 	
 	@PostMapping("/flow")
-	public ResponseEntity<?> addSingleFlow(@RequestBody FlowRequest flow) {
+	public ResponseEntity<?> addSingleFlow(@RequestHeader("api_key") String apiKey, @RequestBody FlowRequest flow) {
+		if (!userService.checkAdminAccessByApiKey(apiKey))
+			return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
+		
 		flowService.add(flow.flow_lvl(), flow.course(), flow.flow(), flow.subgroup());
 		return ResponseEntity.ok().build();
 	}
 	
 	@PostMapping("/flows")
 	@Transactional
-	public ResponseEntity<?> addFlows(@RequestBody List<FlowRequest> flows) {
+	public ResponseEntity<?> addFlows(@RequestHeader("api_key") String apiKey, @RequestBody List<FlowRequest> flows) {
+		if (!userService.checkAdminAccessByApiKey(apiKey))
+			return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
+		
 		for (FlowRequest flow : flows) {
 			flowService.add(flow.flow_lvl(), flow.course(), flow.flow(), flow.subgroup());
 		}
@@ -49,7 +61,10 @@ public class FlowController {
 	}
 	
 	@DeleteMapping("/all")
-	public ResponseEntity<?> deleteAllFlows() {
+	public ResponseEntity<?> deleteAllFlows(@RequestHeader("api_key") String apiKey) {
+		if (!userService.checkAdminAccessByApiKey(apiKey))
+			return ResponseEntity.status(HttpStatusCode.valueOf(401)).build();
+		
 		flowService.deleteAll();
 		return ResponseEntity.ok().build();
 	}
